Consolidate front matter code in a module
This change reorganizes the front matter-related code so it is fully
contained within the Bridgetown::FrontMatter module. That makes it
easier to find the functionality and standardizes the way that
Bridgetown refers to front matter.

It also adds deprecated constant proxies for backwards-compatibility.
It's unclear whether the moved modules were public API or not so the
proxies make it so updating will not break anyone's site without
warning.
