+++ bridgetown-core/lib/bridgetown-core/front_matter.rb
@@ -2,6 +2,10 @@
 
 module Bridgetown
   module FrontMatter
+    autoload :Defaults, "bridgetown-core/front_matter/defaults"
+    autoload :Importer, "bridgetown-core/front_matter/importer"
     autoload :Loaders, "bridgetown-core/front_matter/loaders"
+    autoload :RubyDSL, "bridgetown-core/front_matter/ruby"
+    autoload :RubyFrontMatter, "bridgetown-core/front_matter/ruby"
   end
 end
