+++ bridgetown-core/lib/bridgetown-core/front_matter/loaders.rb
@@ -17,6 +17,14 @@ module Bridgetown
         registry.map { |loader_class| loader_class.new(origin_or_layout) }
       end
 
+      # Determines whether a given file has front matter
+      #
+      # @param path [Pathname, String] the path to the file
+      # @return [Boolean] true if the file has front matter, false otherwise
+      def self.front_matter?(file)
+        registry.any? { |loader_class| loader_class.header?(file) }
+      end
+
       # Registers a new type of front matter loader
       #
       # @param loader_class [Loader::Base]
