+++ bridgetown-core/lib/bridgetown-core/readers/plugin_content_reader.rb
@@ -31,7 +31,7 @@ module Bridgetown
       dir = File.dirname(path.sub("#{content_dir}/", ""))
       name = File.basename(path)
 
-      @content_files << if Utils.has_yaml_header?(path) || Utils.has_rbfm_header?(path)
+      @content_files << if FrontMatter::Loaders.front_matter?(path)
                           site.collections.pages.read_resource(path, manifest: manifest)
                         else
                           Bridgetown::StaticFile.new(site, content_dir, "/#{dir}", name)
