+++ bridgetown-core/test/source/src/_posts/2023-06-30-ruby-front-matter.md
@@ -1,0 +1,7 @@
+```ruby
+front_matter do
+  title "Boo"
+end
+```
+
+This is a post with Ruby front matter.
